server/requirements.txt
pyzmq==25.1.0
psycopg2-binary==2.9.9